// src/lib/models/CourseCommentDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CourseCommentDTO = {
    id?: number;
    courseId?: number;
    userId?: string;
    userFullName?: string;
    userPhotoUrl?: string;
    content?: string;
    createdAt?: string;
    updatedAt?: string;
};